Fix test

+++ test/dense.jl
@@ -1350,7 +1350,7 @@ end
 @testset "structure of dense matrices" begin
     # A is neither triangular nor symmetric/Hermitian
     A = [1 im 2; -im 0  3; 2 3 im]
-    @test factorize(A) isa LU{ComplexF64, Matrix{ComplexF64}, Vector{Int64}}
+    @test factorize(A) isa LU{ComplexF64, Matrix{ComplexF64}, Vector{Int}}
     @test !any(LinearAlgebra.getstructure(A))
 end
